# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OLLIE - Outlook Local Language Inference Engine",
    "description": "\n        Intelligent bridge orchestrating communication with the\n        [Universal RAG Service](https://github.com/detti97/universal-rag-service)\n        to provide context-aware AI email suggestions.\n        ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Classical LLM chat",
        "description": "Provide a classical chat interface with history and RAG context.\n\nIf the latest user message contains a meeting request, the reply is\naugmented with real calendar availability and a concrete meeting proposal.",
        "operationId": "postChat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponseSchema"
                }
              }
            }
          },
          "503": {
            "description": "RAG Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/email/suggestion/stream": {
      "post": {
        "tags": [
          "email"
        ],
        "summary": "Generate AI email suggestion with live pipeline progress",
        "description": "Generate a reply suggestion, streaming each pipeline step as it completes.\n\nIf the email contains a meeting request and the calendar is connected, the\npipeline is augmented with real availability, and the final `done` event\ncarries a concrete meeting proposal.",
        "operationId": "streamEmailSuggestion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailSuggestionRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE stream of pipeline steps, ending in a done or error event",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/pdf": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Upload and index a PDF document",
        "operationId": "upload_pdf_knowledge_pdf_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_pdf_knowledge_pdf_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeUploadResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/search": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Search in the knowledge base",
        "operationId": "search_knowledge_knowledge_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSearchResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/knowledge/documents/{filename}": {
      "delete": {
        "tags": [
          "knowledge"
        ],
        "summary": "Delete a document from the knowledge base",
        "operationId": "delete_document_knowledge_documents__filename__delete",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Document Knowledge Documents  Filename  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/auth/login": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Get the Microsoft login URL to connect the calendar",
        "operationId": "getCalendarAuthLogin",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUrlSchema"
                }
              }
            }
          },
          "503": {
            "description": "Graph not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/auth/callback": {
      "post": {
        "tags": [
          "calendar"
        ],
        "summary": "Exchange an OAuth authorization code for Graph tokens",
        "operationId": "postCalendarAuthCallback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCallbackRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatusSchema"
                }
              }
            }
          },
          "503": {
            "description": "Token exchange failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/calendar/events": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "List calendar events in a date range",
        "operationId": "getCalendarEvents",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventListSchema"
                }
              }
            }
          },
          "503": {
            "description": "Graph API unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/meeting-times": {
      "post": {
        "tags": [
          "calendar"
        ],
        "summary": "Find meeting times that work for all given attendees",
        "description": "Find slots where every given attendee (plus the signed-in user) is free.\n\nOnly works for attendees within the same Microsoft 365 tenant, since\nMicrosoft Graph has no visibility into external/private calendars.",
        "operationId": "postCalendarMeetingTimes",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindMeetingTimesRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingTimeSuggestionListSchema"
                }
              }
            }
          },
          "503": {
            "description": "Graph API unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/ics/status": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Check whether an own ICS calendar link is configured",
        "operationId": "getCalendarIcsStatus",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IcsStatusSchema"
                }
              }
            }
          },
          "503": {
            "description": "ICS backend not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/ics/self": {
      "post": {
        "tags": [
          "calendar"
        ],
        "summary": "Set the signed-in user's own published-calendar ICS URL",
        "description": "Validate the given ICS feed URL by fetching it, then store it as \"my calendar\".",
        "operationId": "postCalendarIcsSelf",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSelfIcsUrlRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IcsStatusSchema"
                }
              }
            }
          },
          "422": {
            "description": "URL not reachable/parseable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "503": {
            "description": "ICS backend not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/ics/known": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "List known calendar links of other people",
        "operationId": "getCalendarIcsKnown",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnownCalendarListSchema"
                }
              }
            }
          },
          "503": {
            "description": "ICS backend not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "calendar"
        ],
        "summary": "Save another person's published-calendar ICS URL",
        "operationId": "postCalendarIcsKnown",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetKnownIcsUrlRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnownCalendarListSchema"
                }
              }
            }
          },
          "422": {
            "description": "URL not reachable/parseable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "503": {
            "description": "ICS backend not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/ics/known/{email}": {
      "delete": {
        "tags": [
          "calendar"
        ],
        "summary": "Remove a saved calendar link",
        "operationId": "deleteCalendarIcsKnown",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnownCalendarListSchema"
                }
              }
            }
          },
          "503": {
            "description": "ICS backend not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_upload_pdf_knowledge_pdf_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_pdf_knowledge_pdf_post"
      },
...
      "ChatRequestSchema": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageSchema"
            },
            "type": "array",
            "title": "Messages",
            "description": "The history of the conversation"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "ChatRequestSchema"
      },
      "ChatResponseSchema": {
        "properties": {
          "reply": {
            "type": "string",
            "title": "Reply",
            "description": "The AI-generated reply"
          },
          "meetingProposal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MeetingProposalSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "A concrete meeting suggestion, if the message contained a meeting request."
          }
        },
        "type": "object",
        "required": [
          "reply"
        ],
        "title": "ChatResponseSchema"
      },
      "EmailSuggestionRequestSchema": {
        "properties": {
          "emailContent": {
            "type": "string",
            "minLength": 1,
            "title": "Emailcontent",
            "description": "The full text of the email for which a suggestion should be generated.",
            "examples": [
              "Hello, can we move the meeting tomorrow to 2 PM? Best regards, Max"
            ]
          },
          "attendees": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attendees",
            "description": "Email addresses of the other recipients (To/Cc), used to check everyone's calendar availability for meeting-time suggestions.",
            "examples": [
              [
                "alice@contoso.com"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "emailContent"
        ],
        "title": "EmailSuggestionRequestSchema"
      },
      "ErrorResponseSchema": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "A detailed error message explaining what went wrong.",
            "examples": [
              "RAG Service request failed: connection timeout"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponseSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}